+++ src/index.ts
@@ -1,11 +1,7 @@
 import * as queueApi from '@/api/queue';
-import * as messagesApi from '@/api/messages';
-import * as message from '@/models/message';
 import * as event from '@/models/event';
 
 export {
     queueApi,
-    messagesApi,
-    message,
     event
 };
