--- src/models/message.ts
@@ -1,12 +1,0 @@
-interface Message {
-    messageId: string;
-    chatId: string;
-    senderId: string;
-    userIds: string[];
-    createdAt: number;
-    content: string;
-};
-
-export {
-    Message
-};
